pyproject.toml
[project]
name = "revops-forecast-agent"
version = "0.1.0"
description = "AI Rev Ops agent that runs structured forecast calls with sales leaders, reconciles verbal updates against Salesforce, and produces a quarter-level forecast."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.6",
    "rich>=13.7",
    "typer>=0.12",
]

[project.optional-dependencies]
voice = [
    "pipecat-ai[daily,deepgram,elevenlabs,anthropic]>=0.0.40",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
revops-forecast = "revops_forecast_agent.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
